feat: groupForm>confirm 모달로 재확인 하는 부분 추상화
